refactor: reorganized wick-config structure to consolidate conversion code
