+++ crates/wick/wick-config/src/app_config/resources.rs
@@ -1,5 +1,3 @@
-use crate::v1;
-
 #[derive(Debug, Clone, PartialEq)]
 /// A definition of a Wick Collection with its namespace, how to retrieve or access it and its configuration.
 #[must_use]
@@ -29,15 +27,6 @@ pub enum ResourceDefinition {
   UdpPort(UdpPort),
 }
 
-impl From<v1::ResourceDefinition> for ResourceDefinition {
-  fn from(value: v1::ResourceDefinition) -> Self {
-    match value {
-      v1::ResourceDefinition::TcpPort(v) => Self::TcpPort(v.into()),
-      v1::ResourceDefinition::UdpPort(v) => Self::UdpPort(v.into()),
-    }
-  }
-}
-
 impl From<ResourceDefinition> for TcpPort {
   fn from(value: ResourceDefinition) -> Self {
     match value {
@@ -56,15 +45,6 @@ pub struct TcpPort {
   pub address: String,
 }
 
-impl From<v1::TcpPort> for TcpPort {
-  fn from(value: v1::TcpPort) -> Self {
-    Self {
-      port: value.port,
-      address: value.address,
-    }
-  }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 /// Normalized representation of a UDP port configuration.
 pub struct UdpPort {
@@ -72,13 +52,4 @@ pub struct UdpPort {
   pub port: u16,
   /// The address to bind to.
   pub address: String,
-}
-
-impl From<v1::UdpPort> for UdpPort {
-  fn from(value: v1::UdpPort) -> Self {
-    Self {
-      port: value.port,
-      address: value.address,
-    }
-  }
 }
