--- crates/wick/wick-config/src/parse.rs
@@ -1,6 +1,0 @@
-pub(crate) mod v0;
-pub(crate) mod v1;
-/// The reserved name for components that send static data.
-pub(crate) static SENDER_ID: &str = "core::sender";
-/// The name of SENDER's output port.
-pub(crate) static SENDER_PORT: &str = "output";
