+++ crates/wick/wick-config/src/v0/conversions.rs
@@ -1,0 +1,212 @@
+use std::collections::HashMap;
+use std::time::Duration;
+
+use flow_expression_parser::parse_id;
+use serde_json::Value;
+
+use crate::component_definition::{GrpcUrlComponent, NativeComponent};
+use crate::error::ManifestError;
+use crate::flow_definition::{PortReference, SenderData};
+use crate::host_definition::{HostConfig, MeshConfig};
+use crate::utils::{opt_str_to_ipv4addr, opt_str_to_pathbuf};
+use crate::{
+  parse_default,
+  v0,
+  BoundComponent,
+  ComponentConfiguration,
+  ComponentDefinition,
+  ConnectionDefinition,
+  ConnectionTargetDefinition,
+  FlowOperation,
+  HttpConfig,
+  InstanceReference,
+  ManifestComponent,
+  Permissions,
+  Result,
+  WasmComponent,
+};
+
+impl TryFrom<v0::HostManifest> for ComponentConfiguration {
+  type Error = ManifestError;
+
+  fn try_from(def: v0::HostManifest) -> Result<Self> {
+    let flows: Result<HashMap<String, FlowOperation>> = def
+      .network
+      .schematics
+      .iter()
+      .map(|val| Ok((val.name.clone(), val.try_into()?)))
+      .collect();
+    Ok(ComponentConfiguration {
+      source: None,
+      format: def.format,
+      version: def.version,
+      types: Default::default(),
+      host: def.host.try_into()?,
+      name: def.network.name,
+      import: def
+        .network
+        .collections
+        .iter()
+        .map(|val| {
+          Ok((
+            val.namespace.clone(),
+            BoundComponent::new(val.namespace.clone(), val.try_into()?),
+          ))
+        })
+        .collect::<Result<HashMap<_, _>>>()?,
+      labels: def.network.labels,
+      operations: flows?,
+    })
+  }
+}
+
+impl TryFrom<&crate::v0::CollectionDefinition> for ComponentDefinition {
+  type Error = crate::Error;
+  fn try_from(def: &crate::v0::CollectionDefinition) -> std::result::Result<Self, Self::Error> {
+    let kind = match def.kind {
+      crate::v0::CollectionKind::Native => ComponentDefinition::Native(NativeComponent {}),
+      crate::v0::CollectionKind::GrpcUrl => ComponentDefinition::GrpcUrl(GrpcUrlComponent {
+        url: def.reference.clone(),
+        config: def.data.clone(),
+      }),
+      crate::v0::CollectionKind::WaPC => ComponentDefinition::Wasm(WasmComponent {
+        reference: def.reference.clone(),
+        permissions: json_struct_to_permissions(def.data.get("wasi"))?,
+        config: def.data.clone(),
+      }),
+      crate::v0::CollectionKind::Network => ComponentDefinition::Manifest(ManifestComponent {
+        reference: def.reference.clone(),
+        config: def.data.clone(),
+      }),
+    };
+    Ok(kind)
+  }
+}
+
+fn json_struct_to_permissions(json_perms: Option<&Value>) -> Result<Permissions> {
+  let perms = if let Some(json_perms) = json_perms {
+    serde_json::from_value(json_perms.clone()).map_err(crate::Error::Invalid)?
+  } else {
+    Permissions::default()
+  };
+
+  Ok(perms)
+}
+
+impl TryFrom<&crate::v0::SchematicManifest> for FlowOperation {
+  type Error = ManifestError;
+
+  fn try_from(manifest: &crate::v0::SchematicManifest) -> Result<Self> {
+    let instances: Result<HashMap<String, InstanceReference>> = manifest
+      .instances
+      .iter()
+      .map(|(key, val)| Ok((key.clone(), val.try_into()?)))
+      .collect();
+    let connections: Result<Vec<ConnectionDefinition>> =
+      manifest.connections.iter().map(|def| def.try_into()).collect();
+    Ok(Self {
+      name: manifest.name.clone(),
+      inputs: Default::default(),
+      outputs: Default::default(),
+      instances: instances?,
+      connections: connections?,
+      collections: manifest.collections.clone(),
+      constraints: manifest.constraints.clone().into_iter().collect(),
+    })
+  }
+}
+
+impl TryFrom<crate::v0::ComponentDefinition> for InstanceReference {
+  type Error = ManifestError;
+  fn try_from(def: crate::v0::ComponentDefinition) -> Result<Self> {
+    let (ns, name) = parse_id(&def.id)?;
+    Ok(InstanceReference {
+      namespace: ns.to_owned(),
+      name: name.to_owned(),
+      data: def.data,
+    })
+  }
+}
+
+impl TryFrom<&crate::v0::ComponentDefinition> for InstanceReference {
+  type Error = ManifestError;
+  fn try_from(def: &crate::v0::ComponentDefinition) -> Result<Self> {
+    let (ns, name) = parse_id(&def.id)?;
+    Ok(InstanceReference {
+      namespace: ns.to_owned(),
+      name: name.to_owned(),
+      data: def.data.clone(),
+    })
+  }
+}
+
+impl TryFrom<&crate::v0::ConnectionDefinition> for ConnectionDefinition {
+  type Error = ManifestError;
+
+  fn try_from(def: &crate::v0::ConnectionDefinition) -> Result<Self> {
+    let from: ConnectionTargetDefinition = def.from.clone().try_into()?;
+    let to: ConnectionTargetDefinition = def.to.clone().try_into()?;
+    let default = match &def.default {
+      Some(json_str) => Some(
+        parse_default(json_str).map_err(|e| ManifestError::DefaultsError(from.clone(), to.clone(), e.to_string()))?,
+      ),
+      None => None,
+    };
+    Ok(ConnectionDefinition { from, to, default })
+  }
+}
+
+impl TryFrom<crate::v0::ConnectionTargetDefinition> for ConnectionTargetDefinition {
+  type Error = ManifestError;
+
+  fn try_from(def: crate::v0::ConnectionTargetDefinition) -> Result<Self> {
+    let data = def.data.map(|json| SenderData { inner: json });
+    Ok(ConnectionTargetDefinition {
+      target: PortReference {
+        instance: def.instance,
+        port: def.port,
+      },
+      data,
+    })
+  }
+}
+
+impl TryFrom<crate::v0::HostConfig> for HostConfig {
+  type Error = ManifestError;
+  fn try_from(def: crate::v0::HostConfig) -> Result<Self> {
+    Ok(Self {
+      allow_latest: def.allow_latest,
+      insecure_registries: def.insecure_registries,
+      timeout: Duration::from_millis(def.timeout),
+      id: def.id,
+      mesh: def.mesh.and_then(|v| v.try_into().ok()),
+      rpc: def.rpc.and_then(|v| v.try_into().ok()),
+    })
+  }
+}
+
+impl TryFrom<crate::v0::MeshConfig> for MeshConfig {
+  type Error = ManifestError;
+  fn try_from(def: crate::v0::MeshConfig) -> Result<Self> {
+    Ok(Self {
+      enabled: def.enabled,
+      address: def.address,
+      creds_path: opt_str_to_pathbuf(&def.creds_path)?,
+      token: def.token,
+    })
+  }
+}
+
+impl TryFrom<crate::v0::HttpConfig> for HttpConfig {
+  type Error = ManifestError;
+  fn try_from(def: crate::v0::HttpConfig) -> Result<Self> {
+    Ok(Self {
+      enabled: def.enabled,
+      port: def.port,
+      address: opt_str_to_ipv4addr(&def.address)?,
+      pem: opt_str_to_pathbuf(&def.pem)?,
+      key: opt_str_to_pathbuf(&def.key)?,
+      ca: opt_str_to_pathbuf(&def.ca)?,
+    })
+  }
+}
